Move TryAsyncResolve back to GraphQL project
