+++ src/GraphQL/ResolveFieldContextExtensions.cs
@@ -2,6 +2,7 @@ using GraphQL.Subscription;
 using GraphQL.Types;
 using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace GraphQL
 {
@@ -56,6 +57,35 @@ namespace GraphQL
                 return typedContext;
 
             return new ResolveEventStreamContext<T>(context);
+        }
+
+        public static Task<object> TryAsyncResolve(this IResolveFieldContext context, Func<IResolveFieldContext, Task<object>> resolve, Func<ExecutionErrors, Task<object>> error = null)
+        {
+            return TryAsyncResolve<object>(context, resolve, error);
+        }
+
+        public static async Task<TResult> TryAsyncResolve<TResult>(this IResolveFieldContext context, Func<IResolveFieldContext, Task<TResult>> resolve, Func<ExecutionErrors, Task<TResult>> error = null)
+        {
+            try
+            {
+                return await resolve(context).ConfigureAwait(false);
+            }
+            catch (Exception ex)
+            {
+                if (error == null)
+                {
+                    var er = new ExecutionError(ex.Message, ex);
+                    er.AddLocation(context.FieldAst, context.Document);
+                    er.Path = context.Path;
+                    context.Errors.Add(er);
+                    return default;
+                }
+                else
+                {
+                    var result = error(context.Errors);
+                    return result == null ? default : await result.ConfigureAwait(false);
+                }
+            }
         }
     }
 }
